+++ src/nespresso/api/app.py
@@ -42,3 +42,5 @@ class Tags(Enum):
 
 
 app.include_router(nes_user.router, prefix="/user", tags=[Tags.users])
+
+# TODO: Protect /docs with FastAPI